Current sensor added and steering control working

+++ include/motor_driver.hpp
@@ -40,7 +40,7 @@ public:
 
     bool motorDriverIsFaulty();
 
-    uint16_t checkMotorCurrentDraw();
+    float checkMotorCurrentDraw();
 
     void turnOffMotor();
